requirements.txt
fastapi==0.104.1
uvicorn[standard]
python-multipart
jinja2
requests
markdown
playwright
openai
python-dotenv
matplotlib
plotly
kaleido